feat(Consumo de APis): :sparkles: Ajuste en el worker
Se realizan ajustes en el worker para poder tener la funcion de buscar

+++ components/body.js
@@ -1,14 +1,36 @@
 export default {
+    
     showBody(){
+
         const worker = new Worker('./storage/wsPokemon.js')
         worker.postMessage({type: 'showPokemon'});
         worker.onmessage = function(event) {
-            document.querySelector("#pokemones").insertAdjacentHTML("beforeend",event.data)
-            worker.terminate();
-        };        
+        document.querySelector("#pokemones").insertAdjacentHTML("beforeend",event.data)
+        worker.terminate();
+        }; 
 
-        document.querySelector("#siguiente").addEventListener("click",()=>{
-            document.querySelector("#pokemones").innerHTML="<h2>Siguiente</h2>"
+        document.querySelector("#buscador").addEventListener("click",()=>{
+            let val= document.querySelector("#txt-busqueda").value;
+            const worker = new Worker('./storage/wsPokemon.js')
+            worker.postMessage({type: 'findPokemon', data:val});
+            worker.onmessage = function(e) {
+                document.querySelector("#pokemones").innerHTML= ""+e.data
+                worker.terminate();
+            };      
+            
         })
+        /* document.querySelector("#txt-busqueda").addEventListener("change",(e)=>{
+
+            if(e.target.value===""){
+                    const worker = new Worker('./storage/wsPokemon.js')
+                    worker.postMessage({type: 'showPokemon'});
+                    worker.onmessage = function(event) {
+                    document.querySelector("#pokemones").innerHTML=event.data
+                    worker.terminate();
+                    }
+    
+                
+            }
+        }) */
     }
 }
